Auto-forming Forms!👀
NOT TESTED YET. Created BaseForm, which generates a form dynamically from a JSON schema. It supports TextFields, Checkboxes, Sliders, and more. A yup schema is also generated from the JSON, which allows for validation and easy change handling. This component will likely not be used to generate the main forms of the website (though that is a possibility), but rather for filling out a Routine's inputs.