Reduced environment variables
Had a lot of extra environment variables to make the project easily configurable. This made a lot of sense in the template repo Vrooli is forked from, but not needed anymore. Plus, I ran into an issue where React does not have REAC_APP_SERVER_LOCATION for some reason. With all that in mind, I decided to reduce the number of variables needed, and removed the need for REACT_APP_SERVER_LOCATION in the frontend