✨Fancy logs✨
Server logs are now logged to files so they can be studied later. They include a timestamp and a code. The first part of the code signifies where the log is from, and the second part is a random string for analyzing the exact logs from an error a user ran into (though the code is not actually passed to the frontend yet, so more work will have to be done)