Created InputOutputEdge
Created InputOutputEdge, so inputs and outputs can be connected with their own edges (instead of one fake edge). Might delete this for now, but may come back to this approach later on to add conditions to inputs/outputs. For example, maybe some inputs only show up depending on the choices earlier on. This mechanic is seen in many forms, so it would be useful to add to Vrooli