API updated to match db migration
Updated API to match new db migration, which created support for private data, versioning, storing user input, and improved roles. Server code related to the changes has been commented out, and will be reimplemented in the next few updates. An interesting thing to note is the change from "role" to "permissions" when querying objects. This allows for more access control options, such as having the ability to edit but not delete an object. Later on, it would be cool to be able to assign access control to individual routines/projects/etc. in an organization