Faster development build
Dockerfiles for development previously installed yarn. Since these are always installed on development, there's no need to do this in docker. Instead, node_modules has been added as a volume. This slightly increases the amount of files in the final containers, but memory usage actually goes down a bit. Plus the Dockerfiles are now simpler, and the startup time is quicker