Search history
Added search history to AutocompleteSearchBar, which is unique for every search bar. History options are shown first, with option to delete. They don't show up all the time depending on the search string, but it's a work in progress. Search results also show the object type's icon instead of name, which allows for much more of the title to be shown on small screens