Improved querying
In addition to the existing mutationWrapper for Apollo, now we have queryWrapper and documentNodeWrapper. All three use a helper function to reduce code duplication. The first two are useful most of the time (when you can create a hook). The third is useful when you cannot create a hook (outside of a component), such as for quickActions