type back to __typename
Previously, the "type" field was added to every GraphQL object because Apollo's generated __typename field is undefined. I found a much easier - and obvious in hindsight - way. The generated __typename field is now changed to required using sed, because Apollo always provides it. graphql/code-gen generates it as optional because other GraphQL packages may not provide it