Scheduling setup p1
Previously, the only cron jobs were for calculating object and site stats. Now a folder called "schedules" has been made. It holds the stats cron jobs, and also 2 new folders. One is for count jobs, which check for correct values for score, bookmarks, and other db counts. The second is for events. Since the way events are calculated is not easy to query, we must cache upcoming event times. Then we can use this cache to trigger push notification for event reminders, and also to trigger automatic routines