Initial theme not loaded FIXED
Fixed a very annoying issue where the first time you load the production build, the MUI theme would not work. The issue was that the cookie fallback logic didn't work, and cookies are not set on first load. This caused the font size to be an incorrect value (undefined instead of a number), which createTheme doesn't like.