Embeddings search query approach
Added comment showing an approach for how to implement an embeddings query that factors in stars. I used ChatGPT to come up with this. According to it, this shouldn't cause too many performance issues until tables get around the 1,000,000 rows mark. Stars can be swapped for score, bookmarks, etc. Should also create a number field that stores some sort of reputation (based on reports and if the object is trying to game the embeddings search through repeated words and excessive buzzwords), and factor that in