endpoint/method pair generation p1
Updated `gqlSelect.ts` to read REST endpoints defined in server. The goal is to create pairs of endpoint names and method types (e.g. "GET", "POST"), so these can be used in the UI instead of defining them manually (which can be prone to errors). Right now this gets the endpoints and methods, but doesn't yet name or save them