gqlSelects memory improvements
Updated gqlSelects.ts script, which is used to generate endpoints for the UI to use the API, to use less memory. In the endpoint/method pairs section, it now calculates each file individually in a single loop. Previously, it looped through once to store all of the information in an object, then looped again to do more calculations