Removed Blockfrost
We used Blockfrost to query ADA handles tied to Cardano wallets, so that users could own their handle and transfer it to other applications. Blockfrost has unfortunately shut down, so we had to remove this feature. We may bring it back in the future, but more likely will switch to normal handles