Moved route logic
Route logic moved from shared package to ui. It was originally shared with docs, but now that's not needed, so having it in shared makes no sense anymore. Also, I've been running into a problem where there are multiple versions of React, which breaks the application. Removing the need for it in the shared package should help