GQL to JSON test 1 - ts-json-schema-generator
Demonstrated the ability for the `ts-json-schema-generator` package to convert a typescript type to a JSON schema object. The goal is to convert these types to JSON, so that the database seed can upsert main object types to standards, for Valyxa and other autonomous agents to use.

Some issues with this package:
- Did not get it to convert the whole `graphqlTypes.ts` file, because of some error
- Only supports JSON Schema. Would have to modify the output to create an OpenAPI specification, which we need to become a ChatGPT plugin