chatMessage pre updates
Pre-shaper for chat messages now (untested) supports finding chat information from the query directly, in addition to what's already stored in the database. This is accomplished through the new inputs tree we've been implementing in the last few commits. Additionally, it checks to make sure that you are allowed to create the message, and that bots you're inviting are either public or ones you created. Unfortunately, we had to add a lot of complexity to accomplish this.