Profanity check omitted for private objects
Private objects now skip the profanity check, since it's a little silly limiting what people can put in private notes and reminders, among other things. This does mean that someone could add profanity to a private object and set it to public later. Instead of worry about this scenario, we'll rely on the reporting and reputation system for now.