getParentInfo
Introduced `getParentInfo` function parameter for `validate.isPublic`. This is useful for finding permissions data when an object is being created, as sometimes it will contain the necessary data, but oftentimes it will not. This happens when the data needing authenticate was the child of a parent it checks against for authentication.

For example, let's say you created a new Node object with NodeRoutineList data. When authenticating NodeRoutineList, we're going to need the data from Node. This information can only be accessed by finding the parent's input from the `inputsById` map