Automatic SERVER_LOCATION🫡
Previously, we set the SERVER_LOCATION environment variable in the .env file. This is used for, among other things, setting CORS correctly. This made testing productions builds annoying because, if you forgot to change that environment variable, it would not work correctly.

Now, we automatically set SERVER_LOCATION to "local" or "dns" depending on if your IP address matches the SITE_IP environment variable. Additionally, we echo an error if we detect that the SITE_IP does not correspond to the SERVER_URL domain.