CanConnect
UI object shapes have been updated so that relations which can be connected are wrapped in `CanConnect`. This allow us to provide `__connect: boolean` to indicate that an object should be connected instead of created. Previously (which is still the fallback), we only relied on the lack of additional properties to determine when an object is being created