Codemirror real lazy loading p1
Attempting to add real lazy loading to codemirror. Meaning, there should be no languages increasing the size of the main bundle. The approach I'm taking is to copy all of the languages to the public folder, and load them from there instead of node_modules. I didn't get it working yet, but it's a promising approach