moment-timezone lazy-loaded
`moment-timezone`, which is used to calculate schedules, is now lazy-loaded. This reduces the zipped size of the main bundle even further, down to 617KB from 667KB. It is the last remaining thing that's easy to solve relating to reducing the main bundle size. The next largest thing in the bundle is `react-dom`, which isn't going anywhere anytime soon.