"Enter" vs. "Shift+Enter" button
Added button to DashboardView and ChatCrud which allows you to toggle the behavior of the "Enter" button. By default, it submits. When the button is pressed, "Enter" will create a new line instead. This is only available on desktop, since the virtual keyboard on mobile should already have a "Return" button