initializeMap improved error checking
The map of object types to object-specific functions is loaded asynchronously. Since not every GqlModelType has its own file, it was not clear which failures were due to missing files (which is fine) vs. failed imports (which usually mean there is a circular dependency issue). Now we check if the file exists when there is an import failure, and give more helpful logs